+++ src/DependencyInjection/PasswordStrengthExtension.php
@@ -57,6 +57,18 @@ final class PasswordStrengthExtension extends Extension implements PrependExtens
 
     public function prepend(ContainerBuilder $container): void
     {
+        if ($container->hasExtension('framework')) {
+            $container->prependExtensionConfig('framework', [
+                'assets' => [
+                    'packages' => [
+                        Configuration::ALIAS => [
+                            'base_path' => '/bundles/passwordstrength',
+                        ],
+                    ],
+                ],
+            ]);
+        }
+
         if (!$container->hasExtension('twig')) {
             return;
         }